A better example
This page has always confused me though I got it eventually, but it seems like the documentation is not quite there in trying to explain how these tools work. When referencing it I find myself getting confused again and again, i have to remember what I figured out; especially with the naming convention used here. The target in the example feels like the target element I want TO SCROLL TO (since the name of the function is scrollToElement) though in the example it ends up being the target scroll container. It then seems like the target element is scrolling itself to its own offset on its parent. Maybe this page could use better words to try and disambiguate this. The example I put here is how I understood this and use it which works as expected from my understanding.